+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 2.0.32
 ++++++
+* vm/vmss update: expose `--license-type`
 * vm/vmss: use PATCH for updating identities
 
 2.0.31